Finish the response in nextjs tunnel example
If `res.status().json()` is not called in the `try{}` block the browser thinks that the request is never finished, and will not close the connection. This can eventually lead to the page becoming unresponsive due to the number of open/unfinished requests to  `/api/tunnel`. Explicitly ending the response by returning a status and body/json will allow the browser to finish the response.