Fix tunneling example for Express (#214)
The tunneling example for Express.js that was published a week ago via
PR #211 does not properly handle binary compressed session replay
events. This PR rectifies that and changes references to fastify in the
package.json for this example.

This PR properly fixes
https://github.com/getsentry/sentry-javascript/issues/9047 and
https://github.com/getsentry/sentry/issues/57793 (and a few other
relating issues that probably don't need to be mentioned here).